nested_schemas/tc_p0_py_003/requirements.txt
# Requirements for Employee Registration API
# Circular Person Schema Reference Implementation

Flask==3.0.0
flask-cors==4.0.0
PyYAML==6.0.1
Werkzeug==3.0.1
gunicorn==21.2.0